Sort list by machine identifiers
Previously, the sort order was Name, Vendor, Architecture. However, the Vendor is probably a bit more meaningful for
sorting because it will group the plaforms of a given manufacturer together, whereas the name tends to have more
fluctuation (since there are no repercussions from doing so). The name is also a bit strange in that sometimes the
platform name in the package index (shown in Boards Manager) is different from that in platform.txt (shown in the Boards menu).