Use alternative version syntax for `go-version` input of `actions/setup-go`
Go is only used for installation of a tool in the "Check TSV" GitHub Actions workflow. I just want to use the newest
available version rather than unnecessarily specifying a version, which will then increase the maintenance burden of the
workflow.

Even though it would be possible to accomplish that by omitting the `go-version` input altogether, the
`actions/setup-go` action has a very unintuitive behavior when used in this way: It only adds the `go install` target
path to PATH when you specify a `go-version` input.

So I am forced to provide this input. Previously, I was able to use the "*" wildcard version to specify that the latest
version be used, in accordance with the reference provided by the action's documentation:

https://github.com/npm/node-semver#x-ranges-12x-1x-12-

> Any of X, x, or * may be used to "stand in" for one of the numeric values in the [major, minor, patch] tuple.

Some time after that, the workflow runs mysteriously started failing:

Error: The version: * can't be changed to SemVer notation

I also tried using "x", which is also indicated to be valid by that documentation, but it results in the equivalent
workflow run error.

So I am forced to use an alternative syntax.