Add a "Suppress" field to the spreadsheet
In some cases, a platform may be determined to not have any potential value whatsoever, or perhaps a platform maintainer
would request the removal of an unmaintained platform from public listings.

Removal or omission altogether would likely result in future wasted time or regressions as the potential addition was
later rediscovered and re-evaluated or added back to the registry. A better approach is to leave the entry, but mark it
for suppression (which would be used by a hypothetical published listing generation system).