Add package index source fields
The spreadsheet contains fields that identify the source of the platforms. In addition to being required in the case
where no package index is available, this also allows doing automated scans of the source files and contributing fixes
or enhancements to the platform.

The same might also be useful for the package index files. Previously, the source of these files was not effectively
documented in the spreadsheet. In some cases it might be inferred from a raw.githubusercontent.com URL, but in other
cases the package index is maintained in a repository but published elsewhere.