Adjust spell checker configuration to fix false positive
The codespell tool is used for automated detection of common misspellings in the files of this project.

A recent release of codespell includes an expansion of its misspellings dictionary. One of the additions is "ameba" as a
misspelling of "amoeba". The presence of that word in the project content caused the spell check to fail.

The spelling of the word in the occurrences in this project is intentional so this is a false positive.

The false positive is fixed by adding "ameba" to the ignored words list in the codespell configuration file.