Migrate to Task-based Prettier formatting system
On every push and pull request that affects relevant files, check whether the formatting of supported files is compliant
with the Prettier style.

The updated system makes it easy for contributors to run the same formatting operation locally as will be done by the CI
system, using the Task task runner tool.