Migrate to Task-based editorconfig compliance check system
On every push and pull request that affects relevant files, check whether the formatting of supported files is compliant
with the style defined in the project's .editorconfig file.

The updated system makes it easy for contributors to run the same formatting operation locally as will be done by the CI
system, using the Task task runner tool.

The Go modules system is now used to manage the editorconfig tool dependency versioning and installation.