Add infrastructure for checking Poetry configuration files
Tasks and a GitHub Actions workflow are added to maintain and check for problems in the Poetry Python dependency
management's configuration files.

The tasks provide the following operations:

- Check for invalid data structure of pyproject.toml
- Check for out of sync poetry.lock file

These tasks are ran by the GitHub Actions workflow on any push or pull request that modifies relevant files, and
periodically to check for breakage caused by external changes.