Migrate to Task-based workflow check system
A task is added to validate GitHub Actions workflows.

On every push and pull request that affects relevant files, this task will be executed by the GitHub Actions workflow.

The updated system makes it easy for contributors to run the same checking operation locally as will be done by the CI
system, using the Task task runner tool.