Migrate to Task-based YAML linting
A task is added to lint the project's YAML files.

On every push and pull request that affects relevant files, this task will be executed by the GitHub Actions workflow.

The updated system makes it easy for contributors to run the same checking operation locally as will be done by the CI
system, using the Task task runner tool.