Add `--no-cache` flag to `poetry lock` command in `poetry:sync` task
The `poetry:sync` task is used to ensure the Poetry lockfile is in sync with the Python package dependencies specified
in pyproject.toml.

The "Check Poetry" GitHub Actions workflow runs this task and then checks for a diff in order to catch an out of sync
state between the two files.

I discovered that, even though I didn't get any diff after running the command on my machine, a diff was generated on
the machine of the GitHub Actions runner, causing the check to fail. I hypothesize the cause is the difference of mine
being a Windows machine, while the runner is a Linux machine. It is essential that the results be the same no matter
where the task is ran. I determined this can be accomplished by adding the `--no-cache` flag to the command.