Target specific file paths in `npm:validate` task
Previously the task validated all `package.json` files in the repository. This meant validating many externally
maintained files in the `node_modules` folder, resulting in spurious failures.

Unfortunately the method that would allow making exclusions to the automatic recursive package.json discovery would add
a large amount of complexity to the task (see the `markdown:check-links` task), and make it less friendly to Windows
users. So the approach was chosen of requiring the path to each of the project's npm packages to be configured, as is
already done for the Go module tasks.