Add `--yes` flag to npx command in `ci:validate` task
Due to the need to install an older version of the ajv-cli package to support the GitHub Actions workflow schema without
conflicting with the project scoped ajv-cli dependency, which must use a different version, the npx command must do a
global installation of the package.

By default npx prompts for global installation when a package is not already installed. Adding the `--yes` flag to the
npx command allows the task to be run without interruption.