Configure Dependabot to provide updates for dependencies of site module
The website uses the Go module system to manage the theme dependency. It will be useful to get Dependabot update pull
requests for the dependencies of that module in addition to the one in the root of the repository.

Unfortunately there is an additional obstacle in that the maintainer of the theme dependency is not using the correct
tag format (which I have already reported to them). It is worth at least getting the infrastructure set up in this
repository in preparation for the eventuality of the dependency becoming Go module compliant, and also to prepare for
any other Go module dependencies that might be added in the future.