Adjust Prettier command in task to also format TOML files
Support for formatting TOML files was added to Prettier using a plugin. Prettier only formats the files with the
extensions of the default languages when a directory path is used as the argument.

Previously the prettier command in the `general:format-prettier` task used only a path argument, which caused the TOML
files to still not be formatted even after the installation of the plugin.

The problem is solved by adding a recursive glob for the .toml file extension as an argument in addition to the existing
path argument. The path argument causes Prettier to format all the files of the default languages, as before, and the
added argument causes it to also format all TOML files in the project.