Pin Python version for project to 3.11.x
The newly released Python 3.12 has some incompatibilities with common dependencies (e.g., Dependabot) so it is best to
enforce the use of 3.11 for now.

Although the general approach for the project is to do a full pin of all dependencies, an exception is made for the
dependencies for which automatic version updates via Dependabot are not possible. Python falls under that category so it
is only pinned to the minor version.