Add Go module tool dependency config files to workflow path filters
The project uses several tools that are available as Go modules. The Go module framework is used for the management of
these dependencies, using the technique of a dummy "tools" Go package that imports their packages.

The GitHub Actions workflows that provide automation for the project infrastructure are configured to only run when a
relevant file is modified by a commit or pull request. Previously the files that configure the project's Go module tool
dependencies was not included in these path filters. Those files are relevant to all workflows that use the Go
module-based tools, so they must be included in the paths filters.