Set up SVG development infrastructure
The Prettier XML plugin is installed and configured to provide formatting for SVG and other files in XML format.

The SVGLint tool is installed to check for common problems with the project. A task and GitHub Actions workflow is
provided to allow contributors to easily lint the files locally, and to allow the CI system to do so automatically.