Remove unused fields from site content source front matter
The `page-type` field was added to the front matter during the early experimentation with the Hugo static website
generation tool. The field is not used currently and it is not clear whether such a field will ever be needed so it is
better to remove it for now to avoid unnecessary complexity in the website content source code.