Use custom breadcrumb text
Previously the breadcrumb text was the page name. That could result in overly verbose breadcrumb chains. Since there are
different considerations for a page title compared to a breadcrumb, it will be better to define their texts separately.

The breadcrumb text is now defined by a dedicated field in the front matter of each page.