Add breadcrumbs to custom secondary-header.html partial template
The breadcrumbs feature of the hugo-PaperMod theme is implemented in a partial template. That partial is referenced in
the page templates provided by the theme.

Those page templates are targeted to a blog type of website and so are worthless for this project so this project uses
custom page templates exclusively. It is required to include the breadcrumbs code in the custom templates in order for
the site to have breadcrumbs. This is done as part of a custom "secondary-header.html" partial template, which is
referenced by all the project's individual page templates.