Add underline to link in footer
The standard style for links established by the hugo-PaperMod theme used by the website is for the text to be
underlined.

That style was somehow lost when a custom footer was implemented for the site. It is hereby restored. The approach used
to obtain the underline comes from the theme's CSS code:

https://github.com/adityatelange/hugo-PaperMod/blob/v7.0/assets/css/common/footer.css#L26