Configure Prettier to ignore `go.mod`
After the update, Prettier is now attempting to format the project's go.mod file, but failing. It seems it is
recognizing it as some other language.

The solution is to configure Prettier to ignore the file. It is mostly automatically generated by Go and I don't see
that Prettier actually has support for this language, so there is no downside from ignoring it.